ossa: Generalize how we add blocks to data structures
We have a few different situations now where we want to add a block to
a different data structure, so rather than implementing all of them in
terms of a temporary array as we were before we'll instead introduce an
interface for adding blocks and implement array construction in terms of
it.

This is marginal for the cases we already have, but we'll shortly also be
adding a block work queue in the oana package that will benefit from being
able to append to it without other incidental allocations.