oana: Algorithm for building a dominators table
This makes use of the predecessors table, whose representation is also
adjusted here for a new convention of having each analysis function
produce a different named type. This allows the compiler to help check
that the functions are composed together in a reasonable way and makes
the relationships between the algorithms clearer through the function
signatures.